Training: loss calculation and acc eval

+++ sarcasmData.py
@@ -7,6 +7,6 @@ class SarcasmData(data.Dataset):
     def __init__(self, originals, responses, labels):
         super(SarcasmData, self).__init__()
 
-        self.originals_idxs = torch.ByteTensor(originals)
-        self.responses_idxs = torch.ByteTensor(responses)
-        self.labels = torch.ByteTensor(labels)
+        self.originals_idxs = torch.LongTensor(originals)
+        self.responses_idxs = torch.LongTensor(responses)
+        self.labels = torch.LongTensor(labels)
